Add NodeCluster and AutoPositionUtility for enhanced node management and positioning
- Introduced NodeCluster struct to encapsulate node properties and compatibility calculations.
- Implemented AutoPositionUtility for aligning nodes based on specified criteria.
- Added utility extensions for SCNNode and simd_float3 to facilitate node type identification and transformations.
- Updated project.pbxproj to include new source files and ensure proper build configuration.
