+++ ARL Creator/View/SCNView/SCNViewUpdatePositionRoomContainer.swift
@@ -523,537 +523,3 @@ struct SCNViewUpdatePositionRoomContainer_Previews: PreviewProvider {
         SCNViewUpdatePositionRoomContainer()
     }
 }
-
-struct AutoPositionUtility {
-    // Lista dei tipi di nodo da mantenere
-    static let targetTypes = ["Door", "Opening", "Window"]
-    
-    static func findBestAlignment(
-        from sourceNodes: [SCNNode],
-        to targetNodes: [SCNNode],
-        clusterSize: Int = 3,
-        maxPairs: Int = 1000
-    ) -> (rotationAngle: Float, translation: simd_float3, error: Float) {
-        let filteredSourceNodes = filterNodesByType(nodes: sourceNodes)
-        let filteredTargetNodes = filterNodesByType(nodes: targetNodes)
-        
-        let clusters = findCompatibleClusters(
-            from: filteredSourceNodes,
-            and: filteredTargetNodes,
-            clusterSize: clusterSize,
-            maxPairs: maxPairs
-        )
-        
-        var minError = Float(-1);
-        var bestRotationAngle: Float = 0.0;
-        var bestTranslation: simd_float3 = simd_float3(0, 0, 0);
-        
-        // Itera su tutte le combinazioni
-        for (sourceCluster, targetCluster, compatibilityError) in clusters {
-            let (rotationAngle, translation) = computeTransformation(from: sourceCluster, to: targetCluster)
-            
-            // Calcola l'errore di trasformazione
-            let transformationError = computeTransformationError(
-                for: sourceCluster,
-                rotationAngle: rotationAngle,
-                translation: translation,
-                comparedTo: targetCluster
-            )
-            
-            // Calcola l'errore totale
-            let error = compatibilityError + transformationError
-            
-            if (error < minError || minError == Float(-1)) {
-                minError = error
-                
-                bestRotationAngle = rotationAngle;
-                bestTranslation = translation;
-            }
-            
-            print("Error: \(error)")
-            print("Min error: \(minError)")
-            print()
-            print("rotationAngle: \(rotationAngle) (\(rotationAngle * 180 / .pi)°)")
-            print("translation: \(translation)")
-            print()
-        }
-        
-        return (bestRotationAngle, bestTranslation, minError)
-    }
-    
-    private static func filterNodesByType(nodes: [SCNNode]) -> [SCNNode] {
-        return nodes.filter { node in targetTypes.contains { targetType in node.type == targetType }}
-    }
-    
-    // Funzione principale per trovare le coppie di cluster compatibili
-    private static func findCompatibleClusters(
-        from nodes1: [SCNNode],
-        and nodes2: [SCNNode],
-        clusterSize: Int = 3,
-        maxPairs: Int = 1000,
-        maxNodes1: Int = 20,
-        maxNodes2: Int = 40,
-        maxClusters1: Int = 1000,
-        maxClusters2: Int = 10000
-    ) -> [(NodeCluster, NodeCluster, Float)] {
-        // Controllo che ci siano abbastanza nodi
-        guard
-            nodes1.count >= clusterSize,
-            nodes2.count >= clusterSize
-        else {
-            return []
-        }
-        
-        // Limito il numero di nodi da combinare per ottimizzazione
-        let sampledNodes1 = maxNodes1 >= nodes1.count ? nodes1 : Array(nodes1.shuffled().prefix(maxNodes1))
-        
-        // Limito il numero di nodi da combinare per ottimizzazione
-        let sampledNodes2 = maxNodes2 >= nodes2.count ? nodes2 : Array(nodes2.shuffled().prefix(maxNodes2))
-        
-        // Genero tutti i possibili cluster dalla prima lista
-        let clusters1 = sampledNodes1.combinations(taking: clusterSize).map { NodeCluster(nodes: $0) }
-        
-        // Genero tutti i possibili cluster dalla seconda lista
-        let clusters2 = sampledNodes2.combinations(taking: clusterSize).map { NodeCluster(nodes: $0) }
-        
-        // Array per memorizzare i risultati (cluster1, cluster2, error)
-        var compatibilityErrors: [(NodeCluster, NodeCluster, Float)] = []
-        
-        // Limito il numero di cluster da confrontare per ottimizzazione
-        let sampledClusters1 = maxClusters1 >= clusters1.count ? clusters1 : Array(clusters1.shuffled().prefix(maxClusters1))
-        
-        // Limito il numero di cluster da confrontare per ottimizzazione
-        let sampledClusters2 = maxClusters2 >= clusters2.count ? clusters2 : Array(clusters2.shuffled().prefix(maxClusters2))
-        
-        // Calcolo i punteggi di compatibilità
-        for cluster1 in sampledClusters1 {
-            for cluster2 in sampledClusters2 {
-                let error = cluster1.computeCompatibilityError(with: cluster2)
-                compatibilityErrors.append((cluster1, cluster2, error))
-            }
-        }
-        
-        // Ordino in base al punteggio (più basso = più compatibile)
-        compatibilityErrors.sort { $0.2 < $1.2 }
-        
-        // Prendo le migliori coppie
-        let bestPairs = Array(compatibilityErrors.prefix(maxPairs))
-        
-        return bestPairs
-    }
-    
-    // Funzione per calcolare la rotazione attorno all'asse y e la traslazione x,z
-    private static func computeTransformation(
-        from sourceCluster: NodeCluster,
-        to targetCluster: NodeCluster
-    ) -> (rotationAngle: Float, translation: simd_float3) {
-        guard sourceCluster.size == targetCluster.size, sourceCluster.size >= 3 else {
-            fatalError("Sono necessari cluster di almeno 3 nodi")
-        }
-        
-        // 1. Recupera punti e centroidi
-        let sourcePoints = sourceCluster.points
-        let targetPoints = targetCluster.points
-        
-        let sourceCentroid = sourceCluster.centroid
-        let targetCentroid = targetCluster.centroid
-        
-        // 2. Sottrai i centroidi dai punti
-        let centeredSourcePoints = sourcePoints.map { $0 - sourceCentroid }
-        let centeredTargetPoints = targetPoints.map { $0 - targetCentroid }
-        
-        // 3. Trova l'angolo di rotazione ottimale attorno all'asse y
-        var numerator: Float = 0
-        var denominator: Float = 0
-        
-        for i in 0..<sourcePoints.count {
-            let source = centeredSourcePoints[i]
-            let target = centeredTargetPoints[i]
-            
-            // Per rotazione attorno all'asse y: x' = x*cos(θ) + z*sin(θ), z' = -x*sin(θ) + z*cos(θ)
-            numerator += source.z * target.x - source.x * target.z
-            denominator += source.x * target.x + source.z * target.z
-        }
-        
-        let rotationAngle = atan2(numerator, denominator)
-        
-        // 4. Calcola la matrice di rotazione
-        let cosTheta = cos(rotationAngle)
-        let sinTheta = sin(rotationAngle)
-        
-        let rotationMatrix = simd_float3x3(
-            simd_float3(cosTheta, 0, sinTheta),
-            simd_float3(0, 1, 0),
-            simd_float3(-sinTheta, 0, cosTheta)
-        )
-        
-        // 5. Ruota i punti centrati e calcola la traslazione
-        let rotatedSourcePoints = sourcePoints.map { rotationMatrix.transpose * $0 }
-        
-        // La traslazione è la differenza tra il centroide del target e il centroide del source ruotato
-        let rotatedSourceCentroid = rotatedSourcePoints.getCentroid()
-        let translation = targetCentroid - rotatedSourceCentroid
-        
-        return (rotationAngle, translation)
-    }
-    
-    // Funzione per calcolare l'errore di allineamento
-    private static func computeTransformationError(
-        for sourceCluster: NodeCluster,
-        rotationAngle: Float,
-        translation: simd_float3,
-        comparedTo targetCluster: NodeCluster
-    ) -> Float {
-        guard sourceCluster.size == targetCluster.size else {
-            fatalError("Clusters must have the same number of nodes")
-        }
-        
-        let sourcePoints = sourceCluster.points
-        let targetPoints = targetCluster.points
-        
-        return sourcePoints.enumerated().reduce(0.0) { (totalError, element) in
-            let i = element.offset
-            let transformedPoint = sourcePoints[i].transformXZ(rotationAngle: rotationAngle, translation: translation)
-            let error = simd_length(transformedPoint - targetPoints[i])
-            
-            print("A: \(sourcePoints[i])")
-            print("transformedPoint: \(transformedPoint)")
-            print("B: \(targetPoints[i])")
-            print()
-            
-            return totalError + pow(error, 2)
-        }
-    }
-}
-
-struct NodeCluster {
-    let nodes: [SCNNode]
-    let points: [simd_float3]
-    let centroid: simd_float3
-    let distanceMatrix: [[Float]]
-    let relativeHeights: [Float]
-    let dimensions: [(width: Float, height: Float, depth: Float)]
-    let types: [String?]
-    
-    var size: Int {
-        return self.nodes.count
-    }
-    
-    init(nodes: [SCNNode]) {
-        guard nodes.count >= 3 else {
-            fatalError("Sono necessari almeno 3 nodi")
-        }
-        
-        // Calcolo centroide
-        let centroid = nodes.simdWorldPositions.getCentroid()
-        
-        // Ordinamento nodi
-        let sortedNodes = nodes.sorted { distance($0.simdWorldPosition, centroid) < distance($1.simdWorldPosition, centroid) }
-        
-        // Calcolo punti
-        let points = sortedNodes.simdWorldPositions
-        
-        // Calcolo distanze
-        let distanceMatrix = (sortedNodes.simdWorldPositions + [centroid]).getDistanceMatrix()
-        
-        // Calcolo altezze relative
-        let relativeHeights = sortedNodes.simdWorldPositions.getRelativeHeights()
-        
-        // Calcolo dimensioni
-        let dimensions = sortedNodes.map { $0.boundingBoxDim }
-        
-        // Calcolo tipi
-        let types = sortedNodes.map { $0.type }
-        
-        self.nodes = sortedNodes
-        self.centroid = centroid
-        self.points = points
-        self.distanceMatrix = distanceMatrix
-        self.relativeHeights = relativeHeights
-        self.dimensions = dimensions
-        self.types = types
-    }
-    
-    func computeCompatibilityError(
-        with other: NodeCluster,
-        threshold: Float = 0.1,
-        distanceWeight: Float = 1.0,
-        relativeHeightWeight: Float = 1.0,
-        dimensionWeight: Float = 1.0,
-        typeMismatchWeight: Float = 1.0,
-        typeDiversityWeight: Float = 1.0
-    ) -> Float {
-        let distanceMSE = self.computeDistanceMSE(with: other, threshold: threshold)
-        let relativeHeightMSE = self.computeRelativeHeightMSE(with: other, threshold: threshold)
-        let dimensionMSE = self.computeDimensionMSE(with: other, threshold: threshold)
-        let typeMismatchMSE = self.computeTypeMismatchMSE(with: other, threshold: threshold)
-        let typeDiversityMSE = self.computeTypeDiversityMSE(with: other, threshold: threshold)
-        
-        return (
-            distanceWeight * distanceMSE +
-            relativeHeightWeight * relativeHeightMSE +
-            dimensionWeight * dimensionMSE +
-            typeMismatchWeight * typeMismatchMSE +
-            typeDiversityWeight * typeDiversityMSE
-        )
-    }
-    
-    func computeDistanceMSE(with other: NodeCluster, threshold: Float = 0.1) -> Float {
-        guard
-            self.distanceMatrix.count > 0,
-            self.distanceMatrix[0].count > 0,
-            self.distanceMatrix.count == other.distanceMatrix.count,
-            self.distanceMatrix[0].count == other.distanceMatrix[0].count
-        else {
-            fatalError("The two distance matrices must have the same non-zero dimensions.")
-        }
-
-        var mse: Float = 0.0
-        
-        for (i, row) in self.distanceMatrix.enumerated() {
-            for (j, _) in row.enumerated() where j > i {
-                let d1 = self.distanceMatrix[i][j]
-                let d2 = other.distanceMatrix[i][j]
-                
-                let diff = abs(d1 - d2)
-                mse += diff < threshold ? 0 : pow(diff, 2)
-            }
-        }
-
-        return mse
-    }
-    
-    func computeRelativeHeightMSE(with other: NodeCluster, threshold: Float = 0.1) -> Float {
-        guard
-            self.relativeHeights.count > 0,
-            self.relativeHeights.count == other.relativeHeights.count
-        else {
-            fatalError("The two relative height arrays must have the same non-zero dimension.")
-        }
-
-        var mse: Float = 0.0
-        
-        for (i, _) in self.relativeHeights.enumerated() {
-            let h1 = self.relativeHeights[i]
-            let h2 = other.relativeHeights[i]
-            
-            let diff = abs(h1 - h2)
-            mse += diff < threshold ? 0 : pow(diff, 2)
-        }
-
-        return mse
-    }
-    
-    func computeDimensionMSE(with other: NodeCluster, threshold: Float = 0.1) -> Float {
-        guard
-            self.dimensions.count > 0,
-            self.dimensions.count == other.dimensions.count
-        else {
-            fatalError("The two dimension arrays must have the same non-zero dimension.")
-        }
-
-        var mse: Float = 0.0
-        
-        for (i, _) in self.dimensions.enumerated() {
-            let (w1, h1, d1) = self.dimensions[i]
-            let (w2, h2, d2) = other.dimensions[i]
-            
-            let wDiff = abs(w1 - w2)
-            mse += wDiff < threshold ? 0 : pow(wDiff, 2)
-            
-            let hDiff = abs(h1 - h2)
-            mse += hDiff < threshold ? 0 : pow(hDiff, 2)
-            
-            let dDiff = abs(d1 - d2)
-            mse += dDiff < threshold ? 0 : pow(dDiff, 2)
-        }
-
-        return mse
-    }
-    
-    func computeTypeMismatchMSE(with other: NodeCluster, threshold: Float = 0.1) -> Float {
-        guard
-            self.types.count > 0,
-            self.types.count == other.types.count
-        else {
-            fatalError("The two type arrays must have the same non-zero dimension.")
-        }
-        
-        var totalError: Float = 0.0
-        
-        for (i, _) in self.types.enumerated() {
-            let selfType = self.types[i] ?? ""
-            let otherType = other.types[i] ?? ""
-            
-            // Calcola un errore binario: 0 se i tipi sono uguali, 1 se sono diversi
-            let typeError: Float = (selfType == otherType) ? 0.0 : 1.0
-            
-            totalError += typeError
-        }
-        
-        // Calcola la media degli errori
-        let mse = self.size > 0 ? totalError / Float(self.size) : 1.0
-        
-        // Applica la soglia, se necessario
-        return mse < threshold ? 0.0 : mse
-    }
-    
-    func computeTypeDiversityMSE(with other: NodeCluster, threshold: Float = 0.1) -> Float {
-        guard
-            self.types.count > 0,
-            self.types.count == other.types.count
-        else {
-            fatalError("The two type arrays must have the same non-zero dimension.")
-        }
-        
-        // Raccoglie tutti i tipi unici nei due cluster
-        let selfTypes = Set<String?>(self.types)
-        let otherTypes = Set<String?>(other.types)
-        
-        // Calcola l'unione dei tipi unici tra i due cluster
-        let unionTypes = selfTypes.union(otherTypes)
-        
-        // Se non ci sono tipi, restituisci un valore neutro
-        if unionTypes.isEmpty {
-            return 0.0
-        }
-        
-        // Calcola quanti tipi diversi ci sono in totale tra i due cluster
-        let totalUniqueTypes = Float(unionTypes.count)
-        
-        // Calcola il punteggio di diversità normalizzato (0-1)
-        // Più alto è il valore, maggiore è la diversità
-        let diversityScore = totalUniqueTypes / max(Float(selfTypes.count + otherTypes.count), 1.0)
-        
-        // Invertiamo il punteggio per trasformarlo in un "errore"
-        // 1.0 - diversityScore fa sì che alta diversità = basso errore
-        let mse = 1.0 - diversityScore
-        
-        // Applica la soglia, se necessario
-        return mse < threshold ? 0.0 : mse
-    }
-}
-
-extension SCNNode {
-    var type: String? {
-        guard let name = self.name else {
-            return nil
-        }
-        
-        let pattern = "(?:^clone_)+|\\d+$"
-        guard let regex = try? NSRegularExpression(pattern: pattern) else {
-            return nil
-        }
-        
-        let range = NSRange(name.startIndex..<name.endIndex, in: name)
-        return regex.stringByReplacingMatches(in: name, options: [], range: range, withTemplate: "")
-    }
-    
-    var boundingBoxDim: (width: Float, height: Float, depth: Float) {
-        // Ottieni il boundingBox del nodo
-        let (min, max) = self.boundingBox
-        
-        // Calcola la differenza tra i valori max e min per ottenere la dimensione
-        let width = max.x - min.x
-        let height = max.y - min.y
-        let depth = max.z - min.z
-        
-        return (width, height, depth)
-    }
-}
-
-extension simd_float3 {
-    
-    // Funzione per applicare la trasformazione a un punto
-    func transformXZ(rotationAngle: Float = 0, translation: simd_float3 = simd_float3(0, 0, 0)) -> simd_float3 {
-        let cosTheta = cos(rotationAngle)
-        let sinTheta = sin(rotationAngle)
-        
-        // Matrice di rotazione attorno all'asse y
-        let rotationMatrix = simd_float3x3(
-            simd_float3(cosTheta, 0, sinTheta),
-            simd_float3(0, 1, 0),
-            simd_float3(-sinTheta, 0, cosTheta)
-        )
-        
-        // Applica rotazione e poi traslazione
-        let rotated = rotationMatrix.transpose * self
-        return rotated + translation
-    }
-}
-
-extension Array where Element == SCNNode {
-    var simdWorldPositions: [simd_float3] {
-        return self.map { $0.simdWorldPosition }
-    }
-}
-
-extension Array where Element == simd_float3 {
-    
-    // Funzione per calcolare il centroide di un array di punti
-    func getCentroid() -> simd_float3 {
-        var sum = simd_float3(0, 0, 0)
-        for point in self {
-            sum += point
-        }
-        return sum / Float(self.count)
-    }
-    
-    // Metodo per calcolare la matrice delle distanze
-    func getDistanceMatrix() -> [[Float]] {
-        var distanceMatrix: [[Float]] = Array<[Float]>(
-            repeating: Array<Float>(repeating: 0.0, count: self.count),
-            count: self.count
-        )
-        
-        for (i, row) in distanceMatrix.enumerated() {
-            for (j, _) in row.enumerated() where j > i {
-                distanceMatrix[i][j] = simd.distance(self[i], self[j])
-                distanceMatrix[j][i] = distanceMatrix[i][j]
-            }
-        }
-        
-        return distanceMatrix
-    }
-    
-    // Metodo per calcolare le altezze relative
-    func getRelativeHeights() -> [Float] {
-        let minY = self.map { $0.y }.min()!
-        return self.map { $0.y - minY }
-    }
-}
-
-extension Array {
-    func permutations() -> [[Element]] {
-        guard count > 0 else { return [[]] }
-        
-        return indices.flatMap { i -> [[Element]] in
-            var rest = self
-            let element = rest.remove(at: i)
-            return rest.permutations().map { [element] + $0 }
-        }
-    }
-    
-    func combinations(taking k: Int) -> [[Element]] {
-        guard k > 0 else { return [[]] }
-        guard k <= self.count else { return [] }
-
-        if k == self.count {
-            return [self]
-        }
-
-        if k == 1 {
-            return self.map { [$0] }
-        }
-
-        var result: [[Element]] = []
-
-        for (i, element) in self.enumerated() {
-            let remaining = Array(self[(i + 1)...])
-            let subcombinations = remaining.combinations(taking: k - 1)
-            result += subcombinations.map { [element] + $0 }
-        }
-
-        return result
-    }
-}
